# docs/openapi/catalogo.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Catálogo de filmes — API pública",
    "version": "1.0.0",
    "description": "Sessão pelo cookie HttpOnly token. Faça login em /login.html no mesmo navegador ou execute POST /api/auth/login com uma conta local. Não cole o cookie no Authorize. GET /api/auth/me sem sessão retorna 401 e é uma chamada segura para demonstrar Try it out. Papel comum: usuario; administrador: admin. Alterações de dados devem ser testadas somente no ambiente local."
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesma origem do catálogo (local ou publicada)"
    }
  ],
  "components": {
    "schemas": {
      "Mensagem": {
        "type": "object",
        "properties": {
          "mensagem": {
            "type": "string"
          }
        },
        "required": [
          "mensagem"
        ]
      },
...
    },
    "securitySchemes": {
      "sessao": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Cookie HttpOnly criado por POST /api/auth/login. Faça login pelo formulário /login.html no mesmo navegador ou execute o login no Swagger. O navegador envia o cookie automaticamente; não leia nem cole o cookie no Authorize. SameSite=Lax; Secure em produção; duração 8 horas."
      }
    }
  },
  "paths": {
...
    "/api/favoritos": {
      "get": {
        "tags": [
          "Favoritos"
        ],
        "summary": "Listar favoritos próprios",
        "description": "Retorna somente IDs TMDB, na ordem criado_em DESC. Requer sessão autenticada (cookie HttpOnly token).",
        "operationId": "get_api_favoritos",
        "security": [
          {
            "sessao": []
          }
        ],
        "responses": {
          "200": {
            "description": "IDs favoritos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "favoritos": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                      }
                    }
                  },
                  "required": [
                    "favoritos"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado. / Token não informado. / Sessão inválida ou expirada. / Token com usuário inválido. / Usuário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Usuário não autenticado."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Token não informado."
                    }
                  },
                  "2": {
                    "value": {
                      "mensagem": "Sessão inválida ou expirada."
                    }
                  },
                  "3": {
                    "value": {
                      "mensagem": "Token com usuário inválido."
                    }
                  },
                  "4": {
                    "value": {
                      "mensagem": "Usuário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao buscar favoritos. / Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Erro ao buscar favoritos."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Erro interno do servidor."
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Serviço de autenticação indisponível."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/favoritos/{movieId}": {
      "post": {
        "tags": [
          "Favoritos"
        ],
        "summary": "Adicionar favorito",
        "description": "Opera somente na conta autenticada. Sucesso: Filme adicionado aos favoritos! Requer sessão autenticada (cookie HttpOnly token).",
        "operationId": "post_api_favoritos_movieId",
        "security": [
          {
            "sessao": []
          }
        ],
        "responses": {
          "201": {
            "description": "Filme adicionado aos favoritos!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "mensagem": "Filme adicionado aos favoritos!"
                }
              }
            }
          },
          "400": {
            "description": "ID do filme inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "ID do filme inválido."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado. / Token não informado. / Sessão inválida ou expirada. / Token com usuário inválido. / Usuário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Usuário não autenticado."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Token não informado."
                    }
                  },
                  "2": {
                    "value": {
                      "mensagem": "Sessão inválida ou expirada."
                    }
                  },
                  "3": {
                    "value": {
                      "mensagem": "Token com usuário inválido."
                    }
                  },
                  "4": {
                    "value": {
                      "mensagem": "Usuário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Este filme já está nos favoritos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Este filme já está nos favoritos."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao adicionar favorito. / Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Erro ao adicionar favorito."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Erro interno do servidor."
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Serviço de autenticação indisponível."
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "description": "Inteiro positivo.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Favoritos"
        ],
        "summary": "Remover favorito",
        "description": "Opera somente na conta autenticada. Sucesso: Favorito removido com sucesso! Requer sessão autenticada (cookie HttpOnly token).",
        "operationId": "delete_api_favoritos_movieId",
        "security": [
          {
            "sessao": []
          }
        ],
        "responses": {
          "200": {
            "description": "Favorito removido com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "mensagem": "Favorito removido com sucesso!"
                }
              }
            }
          },
          "400": {
            "description": "ID do filme inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "ID do filme inválido."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado. / Token não informado. / Sessão inválida ou expirada. / Token com usuário inválido. / Usuário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Usuário não autenticado."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Token não informado."
                    }
                  },
                  "2": {
                    "value": {
                      "mensagem": "Sessão inválida ou expirada."
                    }
                  },
                  "3": {
                    "value": {
                      "mensagem": "Token com usuário inválido."
                    }
                  },
                  "4": {
                    "value": {
                      "mensagem": "Usuário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Favorito não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Favorito não encontrado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao remover favorito. / Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Erro ao remover favorito."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Erro interno do servidor."
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Serviço de autenticação indisponível."
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "description": "Inteiro positivo.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          }
        ]
      }
    },
...
  },
...
}